requirements.txt
openai-whisper
moviepy
tqdm
